+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.121] - Not released
+### Added
+- Ability to translate the text of comments and posts (if supported by the
+  server). Users can set their language for translation in the preferences (the
+  browser language is used by default). There is a new "Translate to..." item in
+  the "More" menu for comments/posts.
 ### Fixed
 - User can now search for "My feed" in feeds selector